feat(reactivity): add Vue.delete workaround
