+++ src/apis/index.ts
@@ -1,4 +1,5 @@
 export {
+  del,
   isReactive,
   isRef,
   isRaw,
